+++ app/Http/Controllers/Admin/UserController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Http\Requests\UserRequest;
 use App\Models\Painel\Role;
+
 
 class UserController extends Controller
 {
@@ -20,6 +22,10 @@ class UserController extends Controller
 
     public function index()
     {
+        if(Gate::denies('users-view')){
+            abort(403,"Não autorizado!");
+        }
+        
         $totalUsers   = User::count();
 
         \Session::flash('chave','valor');
@@ -34,6 +40,10 @@ class UserController extends Controller
 
     public function role($id)
     {
+        if(Gate::denies('users-edit')){
+            abort(403,"Não autorizado!");
+        }
+
        $user = User::find($id);
        $role = Role::all();
        $paths = [
@@ -47,6 +57,10 @@ class UserController extends Controller
        
     public function roleStore(Request $request, $id)
     {
+        if(Gate::denies('users-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         $user = User::find($id);
         $data = $request->all();
         $role = Role::find($data['role_id']);
@@ -57,6 +71,10 @@ class UserController extends Controller
 
     public function roleDestroy($id, $role_id)
     {
+        if(Gate::denies('users-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         $user = User::find($id);
         $role = Role::find($role_id);
         $user->deleteRole($role);
@@ -68,12 +86,20 @@ class UserController extends Controller
     
     public function create(Request $request)
     {
+        if(Gate::denies('users-create')){
+            abort(403,"Não autorizado!");
+        }
+
         return view('admin.users.create');
     }
 
 
     public function store(UserRequest $request)
     {
+        if(Gate::denies('users-create')){
+            abort(403,"Não autorizado!");
+        }
+       
         $data = $request->only(array_keys($request->rules()));
         User::create($data);
         //\Session::flash('message','Usuário cadastrado com sucesso');
@@ -84,18 +110,30 @@ class UserController extends Controller
 
     public function show(User $user)
     {
+        if(Gate::denies('users-view')){
+            abort(403,"Não autorizado!");
+        }
+        
         return view('admin.users.show', compact('user'));
     }
 
 
     public function edit(User $user)
     {
+        if(Gate::denies('users-edit')){
+            abort(403,"Não autorizado!");
+        }
+        
         return view('admin.users.edit', compact('user'));
     }
 
 
     public function update(UserRequest $request, User $user)
     {
+        if(Gate::denies('users-edit')){
+            abort(403,"Não autorizado!");
+        }
+        
         $data = $request->only(array_keys($request->rules()));
         $user->fill($data);
         $user->save();
@@ -106,6 +144,10 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        if(Gate::denies('users-delete')){
+            abort(403,"Não autorizado!");
+        }
+        
         $user->delete();
         return redirect()->route('users.index')
             ->with('message','Usuário excluído com sucesso');
